Transform pyconfigMasterVertScalePause.yaml to GoLang cluster-syntax (#375)
The following templates in content folder
- deployment-config-1rep-pause-template.json
- deployment-config-2rep-pause-template.json

are merged into 1 file in content folder
golang/deployment-config-pause-template.json

The number of replicas for the dc is controlled by the variable REPLICAS.

The variable IDENTIFIER is generated by regex. So no conflict will
occur when being used as a part of resource names.