Remove timeout and on mb client failure only issue a warning. (#398)
Removing timeout as the real test can take a little longer in some
cases, especially during DNS resolver retries.  The mb tool resolves
the target hosts before running the http(s) tests.  When faced with
a slow DNS server when getaddrinfo() is retried, timeout might kill
the mb client prematurely, also leaving less time for the real
testing.  Timeout command was introduced in the past with badly behaved
clients.  The mb client always exists gracefully after the test
duration, so it is no longer needed.

When mb client exits with a non-zero status, do not fail and try to
copy the mb logs for analysis.  This will help during large-scale
tests, so that possibly other test iterations are not blocked by an
intermittent failure.