Add kraken tool to svt
Kraken is a resiliency tool for OpenShift which can crash a node,
master, etcd and checks if the cluster recovers from the failures.
This is not supported on openshift > 3.9 as the master and etcd as
this tool doesn't take into account that the master services and etcd
are static pods from ocp 3.10.